included hive local db

+++ lib/models/chart_data_model.dart
@@ -1,5 +1,5 @@
 class ChartData {
   ChartData(this.x, this.y);
-  final double x;
-  final double y;
+  final int x;
+  final int y;
 }
